+++ components/Receive/index.jsx
@@ -1,121 +1,50 @@
-import React, { useState } from 'react'
-import { useRouter } from 'next/router'
-import {
-  AccountCard,
-  AccountError,
-  BalanceCard,
-  NetworkSwitcherGlyph,
-  Wrapper,
-  Gutter,
-  Sidebar,
-  Content
-} from '../Shared'
-import { useWalletProvider } from '../../WalletProvider'
-import {
-  LEDGER,
-  CREATE_MNEMONIC,
-  IMPORT_MNEMONIC,
-  ACCOUNT_BATCH_SIZE
-} from '../../constants'
-import {
-  hasLedgerError,
-  reportLedgerConfigError
-} from '../../utils/ledger/reportLedgerConfigError'
-import MsgConfirmer from '../../lib/confirm-message'
-import useUpToDateBalance from '../../lib/update-balance'
-import useWallet from '../../WalletProvider/useWallet'
-import Receive from './Receive'
+import React from 'react'
+import styled from 'styled-components'
+import { func } from 'prop-types'
+import { ADDRESS_PROPTYPE } from '../../customPropTypes'
+import { Address, Box, Glyph, Text, ButtonClose } from '../Shared'
 
-const WalletView = () => {
-  useUpToDateBalance()
-  const wallet = useWallet()
-  const [sending, setSending] = useState(false)
-  const { ledger, walletType, connectLedger } = useWalletProvider()
-  const [uncaughtError, setUncaughtError] = useState(null)
-  const [showLedgerError, setShowLedgerError] = useState(false)
-  const [ledgerBusy, setLedgerBusy] = useState(false)
+const Wrapper = styled(Box).attrs(() => ({
+  display: 'flex',
+  flexDirection: 'column',
+  border: 'none',
+  p: 3,
+  borderRadius: 2,
+  borderWidth: 1,
+  bg: 'background.screen'
+}))``
 
-  const router = useRouter()
-  const onAccountSwitch = () => {
-    const params = new URLSearchParams(router.query)
-    let page = 0
-    if (
-      (wallet && walletType === LEDGER) ||
-      walletType === CREATE_MNEMONIC ||
-      walletType === IMPORT_MNEMONIC
-    ) {
-      page = Math.floor(wallet.path[4] / ACCOUNT_BATCH_SIZE)
-    }
-    params.set('page', page)
-    router.push(`/wallet/accounts?${params.toString()}`)
-  }
-
-  const onShowOnLedger = async () => {
-    setLedgerBusy(true)
-    try {
-      setUncaughtError(null)
-      setShowLedgerError(false)
-      const provider = await connectLedger()
-      if (provider) await provider.wallet.showAddressAndPubKey(wallet.path)
-      else setShowLedgerError(true)
-    } catch (err) {
-      setUncaughtError(err)
-    }
-    setLedgerBusy(false)
-  }
-
+const Receive = ({ address, close }) => {
   return (
-    <>
-      <MsgConfirmer />
-      <Wrapper>
-        <Gutter>
-          <Sidebar>
-            <NetworkSwitcherGlyph />
-            {hasLedgerError(
-              ledger.connectedFailure,
-              ledger.locked,
-              ledger.filecoinAppNotOpen,
-              ledger.replug,
-              ledger.busy,
-              uncaughtError
-            ) && showLedgerError ? (
-              <AccountError
-                onTryAgain={onShowOnLedger}
-                errorMsg={reportLedgerConfigError(
-                  ledger.connectedFailure,
-                  ledger.locked,
-                  ledger.filecoinAppNotOpen,
-                  ledger.replug,
-                  ledger.busy,
-                  uncaughtError
-                )}
-              />
-            ) : (
-              <AccountCard
-                onAccountSwitch={onAccountSwitch}
-                color='purple'
-                alias='Prime'
-                address={wallet.address}
-                isLedgerWallet={walletType === LEDGER}
-                onShowOnLedger={onShowOnLedger}
-                ledgerBusy={ledgerBusy}
-                mb={2}
-              />
-            )}
-            <BalanceCard
-              balance={wallet.balance}
-              disableButtons={sending}
-              onReceive={() => {}}
-              onSend={() => setSending(true)}
-            />
-          </Sidebar>
-          <Content>
-            <Receive address={wallet.address} />
-          </Content>
-        </Gutter>
-      </Wrapper>
-    </>
+    <Wrapper>
+      <Box display='flex' alignItems='center' justifyContent='space-between'>
+        <Box display='flex' alignItems='center'>
+          <Glyph
+            acronym='Rc'
+            color='background.screen'
+            borderColor='core.primary'
+            backgroundColor='core.primary'
+          />
+
+          <Text color='core.primary' ml={2}>
+            Receive Filecoin
+          </Text>
+        </Box>
+        <Box display='flex' alignItems='center'>
+          <ButtonClose ml={2} type='button' onClick={close} />
+        </Box>
+      </Box>
+      <Text my={6}>
+        To receive Filecoin, share your address with the sender
+      </Text>
+      <Address address={address} />
+    </Wrapper>
   )
 }
 
-export default WalletView
+Receive.propTypes = {
+  address: ADDRESS_PROPTYPE,
+  close: func.isRequired
+}
+
+export default Receive
