+++ components/Wallet/views.js
@@ -1,0 +1,3 @@
+export const MESSAGE_HISTORY = 'MESSAGE_HISTORY'
+export const SEND = 'SEND'
+export const RECEIVE = 'RECEIVE'
